move errors.go into mongo package, change an error message
